This commit address 2 items. See VLab issue: for more details https://vlab.ncep.noaa.gov/redmine/issues/64425
1.
Corrects a typo in the srun command option,
underscore should be a dash, and missing an s.

WAS --cpu_bind=core SHOULD BE --cpu-bind=cores

2.
Utilizes the TOTAL_TASKS environment variable, if present, to specify
the number of MPI ranks to use, and adds the -n (--ntasks) option to the srun
command.

This addressed an issue (encountered running HWRF on Theia) where the number
of MPI ranks determined by SLURM did not match the number of TOTAL_TASKS
actually defined by the hwrf configuration and was causing an error when
running WRF.