This commit fixes a bug.
The Bug:
When the Conf2JSON class was added, the ConfigParser Class was added to
the import, from ConfigParser import ConfigParser.

This caused the ConfigParser module to no longer be visible, and the code
would access the ConfigParser class instead. So statements like
ConfigParser.MAX_INTERPOLATION_DEPTH would now error.

The Fix:
Instead using, from ConfigParser import MAX_INTERPOLATION_DEPTH and changed
ConfigParser.MAX_INTERPOLATION_DEPTH to just MAX_INTERPOLATION_DEPTH.

Also in the Conf2JSON class, The reference to the ConfigParser() class
had to be changed to ConfigParser.ConfigParser()