INITIAL Python 3 commit - for METplus
This commit is the "INITIAL" commit of porting this subsetted
produtil package to Python 3.6.0.

This ONLY runs under Python 3.

This commit has ONLY been regression tested against the
METplus code that was also ported to Python 3.
I stress only since METplus doesn't touch all pieces
of produtil so maybe some parts of produtil were
missed being ported, etc.

This commit,
Passed the METplus functional use case tests and Unit Tests
that were being run with the METPlus 2.2 distribution
running under Python 3.6.0

The intent of this commit is to get an initial port
that allows METplus to functionally run so folks can
start working with Python 3. It is expected that more
porting will be needed, as issues are discovered,
and more testing occurs.